Notes were fixed.
