--- src/test/resources/hw3Suite.xml
@@ -1,9 +1,0 @@
-<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd">
-<suite name="Hw3">
-    <test name="SeleniumAutomation">
-        <classes>
-            <class name="hw3.ex1.Exercise1"></class>
-            <class name="hw3.ex2.Exercise2"></class>
-        </classes>
-    </test>
-</suite>
